refactor: 컴포넌트 분리, 코드 컨벤션
