+++ gonggibap/src/app/_components/Sidebar.tsx
@@ -1,199 +1,8 @@
-
-
-// types.ts
-export type Restaurant = {
-  id: number;
-  name: string;
-  rating: number;
-  category: string;
-  distance: string;
-  details: RestaurantDetails;
-};
-
-export type RestaurantDetails = {
-  address: string;
-  openingHours: string;
-  phoneNumber: string;
-  menu: MenuItem[];
-  reviews: Review[];
-};
-
-export type MenuItem = {
-  id: number;
-  name: string;
-  price: number;
-  description: string;
-};
-
-export type Review = {
-  id: number;
-  userName: string;
-  rating: number;
-  content: string;
-  date: string;
-};
-
-// constants.ts
-export const MOBILE_BREAKPOINT = 768;
-export const MOBILE_VIEWS = {
-  LIST: "list",
-  DETAIL: "detail",
-} as const;
-
-export type MobileView = (typeof MOBILE_VIEWS)[keyof typeof MOBILE_VIEWS];
-export type MobilePosition = "peek" | "half" | "full";
-
-// components/RestaurantListView.tsx
-// import React from 'react';
-// import { Restaurant } from '../types';
-
-type RestaurantListViewProps = {
-  restaurants: Restaurant[];
-  onRestaurantSelect: (restaurant: Restaurant) => void;
-  selectedRestaurantId?: number;
-};
-
-export const RestaurantListView: React.FC<RestaurantListViewProps> = ({
-  restaurants,
-  onRestaurantSelect,
-  selectedRestaurantId,
-}) => {
-  return (
-    <div className="space-y-4">
-      {restaurants.map((restaurant) => (
-        <button
-          key={restaurant.id}
-          onClick={() => onRestaurantSelect(restaurant)}
-          className={`w-full text-left p-4 rounded-lg transition-colors
-            ${
-              selectedRestaurantId === restaurant.id
-                ? "bg-gray-700 text-white"
-                : "hover:bg-gray-700"
-            }`}
-        >
-          <div className="flex justify-between items-center">
-            <div>
-              <h3 className="font-bold">{restaurant.name}</h3>
-              <p className="text-sm text-gray-300">{restaurant.category}</p>
-            </div>
-            <div className="text-right">
-              <div className="text-yellow-400">
-                {"⭐".repeat(restaurant.rating)}
-              </div>
-              <p className="text-sm text-gray-300">{restaurant.distance}</p>
-            </div>
-          </div>
-        </button>
-      ))}
-    </div>
-  );
-};
-
-// components/RestaurantDetailView.tsx
-// import React from 'react';
-// import { Restaurant } from '../types';
-
-type RestaurantDetailViewProps = {
-  restaurant: Restaurant;
-  onClose?: () => void; // 웹 닫기 버튼
-  isMobile?: boolean;
-  onBack?: () => void; // 모바일 뒤로가기 버튼
-};
-
-export const RestaurantDetailView: React.FC<RestaurantDetailViewProps> = ({
-  restaurant,
-  onClose,
-  isMobile,
-  onBack,
-}) => {
-  const [isWriting, setIsWriting] = useState<boolean>(false);
-  // 토글하는 함수
-  const onClickWriteReview = () => setIsWriting((prev) => !prev);
-
-  return (
-    <div className="space-y-6">
-      {/* 모바일일 때는 뒤로가기, 웹일 때는 닫기 버튼 */}
-      {isMobile ? (
-        <button
-          onClick={onBack}
-          className="mb-4 px-2 py-1 text-sm rounded bg-gray-700 hover:bg-gray-600"
-        >
-          ← 목록으로
-        </button>
-      ) : (
-        <button
-          onClick={onClose}
-          className="absolute right-0 top-0 p-2 rounded-full hover:bg-gray-600"
-          aria-label="닫기"
-        >
-          ✕
-        </button>
-      )}
-
-      <div>
-        <h2 className="text-xl font-bold mb-2">{restaurant.name}</h2>
-        <div className="space-y-2 text-gray-300">
-          <p>⭐ {restaurant.rating}</p>
-          <p>📍 {restaurant.details.address}</p>
-          <p>🕒 {restaurant.details.openingHours}</p>
-          <p>📞 {restaurant.details.phoneNumber}</p>
-          <button
-            onClick={onClickWriteReview}
-            className={`p-3  rounded-lg ${isMobile ? "bg-gray-700":"bg-gray-800"}`}
-          >
-            리뷰 작성하기
-          </button>
-        </div>
-      </div>
-
-      {isWriting ? (
-        <ReviewForm />
-      ) : (
-        <>
-          <div>
-            <h3 className="text-lg font-bold mb-3">메뉴</h3>
-            <div className="space-y-3">
-              {restaurant.details.menu.map((item) => (
-                <div key={item.id} className="p-3 bg-gray-700 rounded-lg">
-                  <div className="flex justify-between">
-                    <span className="font-medium">{item.name}</span>
-                    <span>{item.price.toLocaleString()}원</span>
-                  </div>
-                  <p className="text-sm text-gray-400">{item.description}</p>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          <div>
-            <h3 className="text-lg font-bold mb-3">리뷰</h3>
-            <div className="space-y-3">
-              {restaurant.details.reviews.map((review) => (
-                <div key={review.id} className="p-3 bg-gray-700 rounded-lg">
-                  <div className="flex justify-between mb-2">
-                    <span className="font-medium">{review.userName}</span>
-                    <span className="text-yellow-400">
-                      {"⭐".repeat(review.rating)}
-                    </span>
-                  </div>
-                  <p className="text-sm mb-1">{review.content}</p>
-                  <p className="text-xs text-gray-400">{review.date}</p>
-                </div>
-              ))}
-            </div>
-          </div>
-        </>
-      )}
-    </div>
-  );
-};
-
-// components/Sidebar.tsx
-import React, { useState, useEffect, TouchEvent } from "react";
-import ReviewForm from "./ReviewForm";
-// import { Restaurant, MobilePosition, MobileView, MOBILE_VIEWS, MOBILE_BREAKPOINT } from '../types';
-// import { RestaurantListView } from './RestaurantListView';
-// import { RestaurantDetailView } from './RestaurantDetailView';
+import { useState, useEffect, TouchEvent } from "react";
+import { MobilePosition, MobileView, Restaurant } from '@/types/sidebar';
+import { RestaurantListView } from "@/app/_components/RestaurantListView";
+import { RestaurantDetailView } from "@/app/_components/RestaurantDetailView";
+import { MOBILE_VIEWS, MOBILE_BREAKPOINT } from "@/constants/sidebar";
 
 type SidebarProps = {
   restaurants: Restaurant[];
