+++ gonggibap/src/constants/sidebar.ts
@@ -1,0 +1,6 @@
+export const MOBILE_BREAKPOINT = 768;
+
+export const MOBILE_VIEWS = {
+  LIST: "list",
+  DETAIL: "detail",
+} as const;
